Move tests of Algorithm in its dedicated file
